20-tests-for-dashboardhandler-go into main

See merge request course/prog2005/prog2005-2025-workspace/olemgl/assignment2!16

+++ utils/structs.go
@@ -82,7 +82,7 @@ type DashboardResponse struct {
 		Capital          string             `json:"capital,omitempty"`
 		Coordinates      map[string]float64 `json:"coordinates,omitempty"`
 		Population       int                `json:"population,omitempty"`
-		Area             int                `json:"area,omitempty"`
+		Area             float64            `json:"area,omitempty"`
 		TargetCurrencies map[string]float64 `json:"targetCurrencies,omitempty"`
 	} `json:"features"`
 	LastRetrieval time.Time `json:"lastRetrieval"`
@@ -93,9 +93,9 @@ type DashboardResponse struct {
  */
 type RestCountriesResponse struct {
 	Capital     []string               `json:"capital"`
-	Coordinates []int                  `json:"latlng"`
+	Coordinates []float64              `json:"latlng"`
 	Population  int                    `json:"population"`
-	Area        int                    `json:"area"`
+	Area        float64                `json:"area"`
 	Currencies  map[string]interface{} `json:"currencies"`
 }
 
@@ -109,6 +109,9 @@ type MetroResponse struct {
 	} `json:"hourly"`
 }
 
+/*
+*	Struct returned from getMetroData()
+ */
 type MetroMeanValues struct {
 	MeanPrecipitation float64
 	MeanTemperature   float64
